+++ CHANGELOG.md
@@ -3,3 +3,4 @@
 * Added "strike" particles to lazors
 * More config screen tweaks
 * Fixed welded Bloques® not rendering properly on Create trains/contraptions
+* New "Vector Suit" built-in resource pack